+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### App Center
 
 * **[Fix]** Remove old support libraries for compatibility with apps without enabled Jetifier tool. 
-* **[Feature]** Add a `setNetworkRequestsAllowed` API to block network requests.
+* **[Feature]** Add a `AppCenter.setNetworkRequestsAllowed(bool)` API to block any network requests without disabling the SDK.
 
 ### App Center Distribute